+++ src/ebic/crud/tomograms.py
@@ -1,11 +1,11 @@
-from fastapi import HTTPException, status
+from fastapi import HTTPException
 from sqlalchemy import func as f
 
-from ..models.response import CtfOut, GenericPlot, TiltAlignmentOut
+from ..models.response import CtfOut, FullMovieWithTilt, GenericPlot
 from ..models.table import CTF, MotionCorrection, Movie, TiltImageAlignment, Tomogram
-from ..utils.database import db
-from ..utils.generic import flatten_join, parse_json_file
-from .path import get_tomogram_auto_proc_attachment
+from ..utils.database import db, paginate
+from ..utils.generic import parse_json_file
+from .movies import get_tomogram_auto_proc_attachment
 
 
 def get_shift_plot(id: int):
@@ -17,56 +17,32 @@ def get_shift_plot(id: int):
     return GenericPlot(items=data)
 
 
-def get_motion_correction(id: int, movie: int = None):
-    raw = (
+def get_motion_correction(limit: int, page: int, tomogramId: int) -> FullMovieWithTilt:
+    query = (
+        db.session.query(MotionCorrection, TiltImageAlignment, CTF, Movie)
+        .filter(TiltImageAlignment.tomogramId == tomogramId)
+        .join(
+            MotionCorrection,
+            MotionCorrection.movieId == TiltImageAlignment.movieId,
+        )
+        .join(Movie, Movie.movieId == TiltImageAlignment.movieId)
+        .join(CTF, CTF.motionCorrectionId == MotionCorrection.motionCorrectionId)
+        .order_by(TiltImageAlignment.refinedTiltAngle)
+    )
+
+    motion = dict(paginate(query, limit, page))
+
+    raw_total = (
         db.session.query(
             f.count(Movie.movieId).label("total"),
         )
         .select_from(Tomogram)
-        .where(Tomogram.tomogramId == id)
+        .where(Tomogram.tomogramId == tomogramId)
         .join(Movie, Movie.dataCollectionId == Tomogram.dataCollectionId)
         .scalar()
     )
 
-    picked = (
-        db.session.query(f.count(TiltImageAlignment.movieId).label("total"))
-        .filter(TiltImageAlignment.tomogramId == id)
-        .scalar()
-    )
-
-    if picked == 0:
-        raise HTTPException(status_code=404, detail="Tomogram has no valid movies")
-
-    if movie is None:
-        movie = picked
-
-    try:
-        query = (
-            db.session.query(MotionCorrection, TiltImageAlignment, CTF, Movie)
-            .filter(TiltImageAlignment.tomogramId == id)
-            .join(
-                MotionCorrection,
-                MotionCorrection.movieId == TiltImageAlignment.movieId,
-            )
-            .join(Movie, Movie.movieId == TiltImageAlignment.movieId)
-            .join(CTF, CTF.motionCorrectionId == MotionCorrection.motionCorrectionId)
-            .order_by(TiltImageAlignment.refinedTiltAngle)
-        )
-
-        query = query.offset(movie - 1).limit(1).first()
-
-    except (TypeError, ValueError):
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Motion correction data does not exist for movie index",
-        )
-
-    return TiltAlignmentOut(
-        drift=parse_json_file(query.MotionCorrection.driftPlotFullPath),
-        total=picked,
-        rawTotal=raw,
-        **flatten_join(query, ["comments"])
-    )
+    return FullMovieWithTilt(**motion, rawTotal=raw_total)
 
 
 def get_ctf(id: int):