+++ src/ebic/routes/sessions.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends
 
 from ..auth import User
-from ..crud import sessions as crud
+from ..crud import groups
 from ..models.response import DataCollectionGroupSummaryOut
 from ..utils.database import Paged
 from ..utils.dependencies import pagination
@@ -10,7 +10,7 @@ router = APIRouter(tags=["Sessions"], prefix="/sessions")
 
 
 @router.get(
-    "/{visitId}/dataGroups", response_model=Paged[DataCollectionGroupSummaryOut]
+    "/{sessionId}/dataGroups", response_model=Paged[DataCollectionGroupSummaryOut]
 )
 def get_collection_groups(
     sessionId: int,
@@ -19,6 +19,6 @@ def get_collection_groups(
     user=Depends(User),
 ):
     """List collection groups belonging to a session"""
-    return crud.get_collection_groups(
+    return groups.get_collection_groups(
         sessionId=sessionId, search=search, user=user, **page
     )