+++ src/ebic/routes/tomograms.py
@@ -2,10 +2,11 @@ from fastapi import APIRouter, Depends
 from fastapi.responses import FileResponse
 
 from ..auth import Permissions
-from ..crud import path, tomograms
-from ..models.response import CtfOut, GenericPlot, TiltAlignmentOut
+from ..crud import movies, tomograms
+from ..models.response import CtfOut, FullMovieWithTilt, GenericPlot
+from ..utils.dependencies import pagination
 
-auth = Permissions("tomogram")
+auth = Permissions.tomogram
 
 router = APIRouter(
     tags=["Tomograms"],
@@ -19,16 +20,18 @@ def get_shift_plot(tomogramId: int = Depends(auth)):
     return tomograms.get_shift_plot(tomogramId)
 
 
-@router.get("/{tomogramId}/motion", response_model=TiltAlignmentOut)
-def get_motion_correction(tomogramId: int = Depends(auth), nth: int = None):
+@router.get("/{tomogramId}/motion", response_model=FullMovieWithTilt)
+def get_motion_correction(
+    tomogramId: int = Depends(auth), page: dict[str, int] = Depends(pagination)
+):
     "Get motion correction data for the given tomogram"
-    return tomograms.get_motion_correction(tomogramId, nth)
+    return tomograms.get_motion_correction(tomogramId=tomogramId, **page)
 
 
 @router.get("/{tomogramId}/centralSlice", response_class=FileResponse)
 def get_slice(tomogramId: int = Depends(auth)):
     """Get tomogram central slice image"""
-    return path.get_tomogram_auto_proc_attachment(tomogramId)
+    return movies.get_tomogram_auto_proc_attachment(tomogramId)
 
 
 @router.get("/{tomogramId}/ctf", response_model=CtfOut)