+++ src/ebic/utils/auth.py
@@ -1,3 +1,7 @@
+from sqlalchemy.orm import Query
+
+from ..auth import User
+from ..models.table import BLSession, Proposal, ProposalHasPerson, SessionHasPerson
 from .config import Config
 
 
@@ -9,14 +13,27 @@ def is_em_staff(perms: list[int]):
     return bool(set(Config.get()["auth"]["read_em_perms"]) & set(perms))
 
 
-def check_admin(func):
-    def wrap(*args, **kwargs):
-        user = args[0]
-        query = args[1]
+def check_session(query: Query, user: User):
+    if is_admin(user.permissions):
+        return query
 
-        if is_admin(user.permissions):
-            return query
+    if is_em_staff(user.permissions):
+        return query.filter(BLSession.beamLineName.like("m__"))
 
-        return func(*args, **kwargs)
+    return query.filter(
+        SessionHasPerson.personId == user.id,
+        SessionHasPerson.sessionId == BLSession.sessionId,
+    )
 
-    return wrap
+
+def check_proposal(query: Query, user: User):
+    if is_admin(user.permissions):
+        return query
+
+    if is_em_staff(user.permissions):
+        return query.join(BLSession).filter(BLSession.beamLineName.like("m__"))
+
+    return query.filter(
+        ProposalHasPerson.personId == user.id,
+        ProposalHasPerson.proposalId == Proposal.proposalId,
+    )